backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-jose[cryptography]
bcrypt
python-multipart
mysql-connector-python
